+++ src/montague/testjson.py
@@ -1,10 +1,9 @@
 from __future__ import absolute_import
 
 import json
-import copy
 import six
 from montague.interfaces import IConfigLoader, IConfigLoaderFactory
-from montague.structs import LoadableConfig, DEFAULT
+from montague.structs import LoadableConfig
 from zope.interface import directlyProvides, implementer
 
 
@@ -27,23 +26,16 @@ class JSONConfigLoader(object):
     def config(self):
         config = {}
         for section, vals in six.iteritems(self._config):
-            if 'main' in vals:
-                vals = copy.deepcopy(vals)
-                vals[DEFAULT] = vals.pop('main')
             config[section] = vals
         return config
 
     def app_config(self, name):
-        if name is DEFAULT:
-            name = 'main'
         # Obviously this will throw a KeyError if the config isn't there.
         # A real implementation would have error handling here.
         config = self._config['application'][name]
         return LoadableConfig.app(name=name, config=config, global_config={})
 
     def server_config(self, name):
-        if name is DEFAULT:
-            name = 'main'
         config = self._config['server'][name]
         return LoadableConfig.server(name=name, config=config, global_config={})