+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ unreleased
 * Enable a ``looponfail`` tox environment.
 * Fix some eval-related bugs in the logging ini conversion.
 * Reincorporate the test JSON config loader, active only during tests.
+* Remove the ``DEFAULT`` sentinal value; we'll use 'main' as the default loadable name, just like grandpa used to do. This is a breaking change.
 
 0.1.5 (2015-05-12)
 -----------------------------------------